Move Guile bindings into pure Scheme

+++ studio/deploy/mac/deploy.sh
@@ -7,14 +7,19 @@ APP=$EXE.app
 VERSION=`git describe --exact-match --tags || echo "($(git rev-parse --abbrev-ref HEAD))"`
 VERSION=`echo $VERSION|sed s:/:-:g`
 
+QT_DIR=$(find /usr/local/Cellar/qt -name "5.1*" -depth 1|head -n1)
+
 cd ../../..
-rm -r build
+rm -rf build
 mkdir build
 cd build
-cmake -DCMAKE_PREFIX_PATH=/usr/local/Cellar/qt/5.9.2 -GNinja -DCMAKE_OSX_DEPLOYMENT_TARGET=10.10 ..
+cmake -GNinja\
+    -DCMAKE_PREFIX_PATH=$QT_DIR \
+    -DLIBFIVE_CCACHE_BUILD=ON \
+    -DCMAKE_OSX_DEPLOYMENT_TARGET=10.12  ..
 rm -rf $APP studio/$APP
 ninja clean
-ninja
+ninja studio/all
 
 # Copy to a new location before modifying, so that the built app doesn't
 # get modified and future builds don't misbehave due to duplicate frameworks
@@ -22,8 +27,8 @@ cp -r studio/$APP $APP
 
 # Pull out framework paths info with otool
 MACDEPLOYQT=`otool -L $APP/Contents/MacOS/$EXE | sed -n -e "s:\(.*\)lib/QtCore.*:\1/bin/macdeployqt:gp"`
-GUILE_SCM=`otool -L $APP/Contents/MacOS/$EXE | sed -n -e "s:lib/libguile.*:share/guile/2.2/:gp"`
-GUILE_CCACHE=`otool -L $APP/Contents/MacOS/$EXE | sed -n -e "s:lib/libguile.*:lib/guile/2.2/ccache/:gp"`
+GUILE_SCM=`otool -L $APP/Contents/MacOS/$EXE | sed -n -e "s:lib/libguile.*:share/guile/3.0/:gp"`
+GUILE_CCACHE=`otool -L $APP/Contents/MacOS/$EXE | sed -n -e "s:lib/libguile.*:lib/guile/3.0/ccache/:gp"`
 
 $MACDEPLOYQT $APP
 
@@ -44,9 +49,29 @@ fix_qt () {
 cd platforms
 fix_qt libqcocoa.dylib
 
-# Delete unused Qt frameworks
+# Delete unused Qt frameworks (wow, there's a lot of them)
 cd ../../Frameworks
-rm -rf QtDeclarative.framework QtMultimedia.framework QtMultimediaWidgets.framework QtPositioning.framework QtQml.framework QtQuick.framework QtScript.framework QtSensors.framework QtSql.framework QtXmlPatterns.framework Qt3DCore.framework Qt3DRender.framework QtLocation.framework QtSerialBus.framework QtSerialPort.framework
+rm -rf \
+    Qt3DCore.framework \
+    Qt3DRender.framework \
+    QtDeclarative.framework \
+    QtLocation.framework \
+    QtMultimedia.framework \
+    QtMultimediaWidgets.framework \
+    QtNetwork.framework \
+    QtPdf.framework \
+    QtPositioning.framework \
+    QtQml.framework \
+    QtQmlModels.framework \
+    QtQuick.framework \
+    QtScript.framework \
+    QtSensors.framework \
+    QtSerialBus.framework \
+    QtSerialPort.framework \
+    QtSql.framework \
+    QtSvg.framework \
+    QtVirtualKeyboard.framework \
+    QtXmlPatterns.framework
 
 # Clean up remaining Qt frameworks
 for LIB in $( ls|sed -n -e "s:\(Qt.*\)\.framework:\1:gp" )
@@ -62,6 +87,18 @@ mkdir -p guile/scm
 mkdir -p guile/ccache
 cp -r $GUILE_SCM guile/scm/
 cp -r $GUILE_CCACHE guile/ccache/
+cp -r ../../../../libfive/bind/guile/libfive guile/scm
+
+# In the Resources directory, find any uncompiled scm files and compile them
+# now for a faster startup, then delete them to reduce bundle size
+cd guile/ccache && find . -name "*.go" | sed "s/\.go//g" | sort > ../../ccache_list
+cd ../scm && find . -name "*.scm" | sed "s/\.scm//g" | sort > ../../scm_list
+cd ../..
+for v in $(comm -2 -3 scm_list ccache_list); do
+    LIBFIVE_FRAMEWORK_DIR=../Frameworks/ guild compile -Lguile/scm -o guile/ccache/$v.go guile/scm/$v.scm
+done
+rm ccache_list scm_list
+rm -r guile/scm
 
 # Update release number in Info.plist
 cd ../../..
@@ -69,7 +106,7 @@ cp ../studio/deploy/mac/Info.plist $APP/Contents/Info.plist
 sed -i "" "s:0\.0\.0:$VERSION:g" $APP/Contents/Info.plist
 
 # Build icon and deploy into bundle
-convert -background none ../studio/deploy/icon/icon.svg -resize 512x512 icon512.png
+inkscape --export-filename=icon512.png ../studio/deploy/icon/icon.svg
 convert icon512.png -resize 256x256 icon256.png
 convert icon512.png -resize 128x128 icon128.png
 convert icon512.png -resize 32x32 icon32.png
@@ -78,13 +115,15 @@ png2icns studio.icns icon512.png icon256.png icon128.png icon32.png icon16.png
 mv studio.icns $APP/Contents/Resources/studio.icns
 rm icon512.png icon256.png icon128.png icon32.png icon16.png
 
-# Create the disk image
-rm -rf deploy $EXE.dmg
-mkdir deploy
-cp ../README.md ./deploy/README.txt
-cp -r ../studio/examples ./deploy/examples
-mv $APP ./deploy
-mkdir deploy/.Trash
-hdiutil create $EXE.dmg -volname "$EXE $VERSION" -srcfolder deploy
-rm -rf deploy
-mv $EXE.dmg ..
+if [ "$1" == "dmg" ]; then
+    # Create the disk image
+    rm -rf deploy $EXE.dmg
+    mkdir deploy
+    cp ../README.md ./deploy/README.txt
+    cp -r ../studio/examples ./deploy/examples
+    mv $APP ./deploy
+    mkdir deploy/.Trash
+    hdiutil create $EXE.dmg -volname "$EXE $VERSION" -srcfolder deploy
+    rm -rf deploy
+    mv $EXE.dmg ..
+fi
